+++ src/OperatorOverloads/Assignment.cpp
@@ -5,22 +5,22 @@ BigInteger& BigInteger::operator=(const BigInteger& rhs) {
 		//std::cout << "=: unequal sizes\n";
 		return *this;
 	}
-	unsigned long long index = this->size;
+	uint64_t index = this->size;
 	while(index--)
 		this->digits[index] = rhs.digits[index];
 	return *this;
 }
 
-BigInteger& BigInteger::operator=(const unsigned long long& rhs) {
+BigInteger& BigInteger::operator=(const uint64_t& rhs) {
 	this->zero();
 	this->digits[0] = rhs;
 	return *this;
 }
 
-unsigned long long BigInteger::operator[](const unsigned long long& index) const {
+uint64_t BigInteger::operator[](const uint64_t& index) const {
 	return digits[index];
 }
 
-unsigned long long& BigInteger::operator[](const unsigned long long& index) {
+uint64_t& BigInteger::operator[](const uint64_t& index) {
 	return digits[index];
 }
